fix(blri): use cross-platform `same-file` operation to optimize possible fs::copy operation
Sometimes when soft/hard links or relativa path exsits, we don't acturally know if the two files are the same only from the file path components. We use `same-file` crate for this purpose.

The original code would cause file being cleaned on Unix, or raising PermissionDenied on Windows, if both the files are the same under different paths.

Signed-off-by: Zhouqi Jiang <luojia@hust.edu.cn>