main: update spi to sdh for loading file from sdcard
This commit replaces the original spi with sdh for two main reasons.
- Adapter board:
  Instead of connecting the appropriate pins for two adapter board,
  just plug in sdcard directly.
  It also improves the stability of loading sdcard.
- Speed:
  Shorten the loading speed of sdcard.

The following are the benchmarks of the two loading speeds, this data
is the average of several consecutive tests, excluding very few values
with large differences, the samples represent normal loading process
in general.

Benchmark Section: SD card and filesystem initialization

+-------------------+-------------+-------------+
| SD card support   |     SPI     |     SDH     |
+-------------------+-------------+-------------+
| Elapsed Cycles    | 936,678,053 | 441,773,813 |
+-------------------+-------------+-------------+
| Elapsed Time (ms) |   1951.41   |   920.36    |
+-------------------+-------------+-------------+

Before is about twice as long as the loading time after the
replacement, which is a significant improvement.

Signed-off-by: DongQing <placebo27@hust.edu.cn>